refactor(bridge): decompose bridge.go into 4 files and extract handleWorkerExit
Split bridge.go (1375 lines) into responsibility-focused files:
- bridge.go (542 lines): struct, lifecycle (Start/Resume/Reset/SwitchWorkDir)
- bridge_worker.go (303 lines): worker creation, crash recovery, agent config
- bridge_forward.go (506 lines): event forwarding, stats, exit handling
- bridge_retry.go (60 lines): LLM auto-retry logic

Extract handleWorkerExit from forwardEvents — the 144-line post-recv exit
handling block is now a standalone method with explicit parameter struct,
improving readability and testability of the event forwarding loop.

Zero logic change: same-package file reorganization with method extraction.

Closes #121 (Phase 1 + 1.5)
