+++ internal/gateway/bridge.go
@@ -2,25 +2,19 @@ package gateway
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log/slog"
-	"os/exec"
-	"runtime/debug"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/hrygo/hotplex/internal/agentconfig"
 	"github.com/hrygo/hotplex/internal/config"
 	"github.com/hrygo/hotplex/internal/eventstore"
 	"github.com/hrygo/hotplex/internal/messaging"
-	"github.com/hrygo/hotplex/internal/metrics"
 	"github.com/hrygo/hotplex/internal/security"
 	"github.com/hrygo/hotplex/internal/session"
 	"github.com/hrygo/hotplex/internal/worker"
-	"github.com/hrygo/hotplex/internal/worker/noop"
 	"github.com/hrygo/hotplex/pkg/aep"
 	"github.com/hrygo/hotplex/pkg/events"
 )
@@ -157,72 +151,6 @@ func (b *Bridge) StartSession(ctx context.Context, id, userID, botID string, wt 
 	return nil
 }
 
-// forwardOpts configures the forwardEvents goroutine behavior.
-type forwardOpts struct {
-	resumed    bool   // true if this goroutine was spawned by ResumeSession
-	workDir    string // workDir to use for resume retry
-	retryDepth int    // number of resume retries attempted (limits to 1)
-	lastInput  string // inherited lastInput from previous retry goroutine; used as fallback when retry worker never receives input
-}
-
-// workerLaunchParams holds the parameters for createAndLaunchWorker.
-type workerLaunchParams struct {
-	ctx         context.Context
-	wt          worker.WorkerType
-	workerInfo  worker.SessionInfo
-	platform    string
-	botID       string
-	forwardOpts *forwardOpts
-}
-
-// workerStartFunc is called after AttachWorker and injectAgentConfig.
-// On startFn failure, the worker is automatically detached.
-type workerStartFunc func(ctx context.Context, w worker.Worker, info worker.SessionInfo) error
-
-// workerAttachErrFunc is called when AttachWorker fails for caller-specific cleanup.
-type workerAttachErrFunc func(w worker.Worker, err error)
-
-// createAndLaunchWorker creates a worker, attaches it, injects config, calls startFn,
-// and launches the forwardEvents goroutine. Returns the worker for post-launch use.
-// On startFn failure, the worker is detached before returning.
-func (b *Bridge) createAndLaunchWorker(params workerLaunchParams, startFn workerStartFunc, attachErrFn workerAttachErrFunc) (worker.Worker, error) {
-	sid := params.workerInfo.SessionID
-	if params.forwardOpts == nil {
-		params.forwardOpts = &forwardOpts{}
-	}
-
-	w, err := b.wf.NewWorker(params.wt)
-	if err != nil {
-		return nil, fmt.Errorf("bridge: create worker: %w", err)
-	}
-
-	if noopw, ok := w.(*noop.Worker); ok {
-		noopw.SetConn(noop.NewConn(sid, params.workerInfo.UserID))
-	}
-
-	if err := b.sm.AttachWorker(sid, w); err != nil {
-		if attachErrFn != nil {
-			attachErrFn(w, err)
-		}
-		return nil, fmt.Errorf("bridge: attach worker: %w", err)
-	}
-
-	b.injectAgentConfig(&params.workerInfo, params.platform, params.botID)
-
-	if err := startFn(params.ctx, w, params.workerInfo); err != nil {
-		b.sm.DetachWorker(sid)
-		return nil, err
-	}
-
-	b.fwdWg.Add(1)
-	go func() {
-		defer b.fwdWg.Done()
-		b.forwardEvents(w, sid, *params.forwardOpts)
-	}()
-
-	return w, nil
-}
-
 // ResumeSession reattaches to an existing session.
 // workDir overrides the stored project directory (used by platform sessions that need a consistent workspace).
 func (b *Bridge) ResumeSession(ctx context.Context, id, workDir string) error {
@@ -354,551 +282,6 @@ func (b *Bridge) resumeWithOpts(ctx context.Context, id, workDir string, opts fo
 // This prevents data races when Hub.Run encodes the clone concurrently with
 // Bridge.forwardEvents encoding the original (e.g., for msgStore.Append).
 var _ = events.Clone // compile-time check that Clone is accessible
-
-func (b *Bridge) persistWorkerSessionID(w worker.Worker, sessionID string) {
-	handler, ok := w.(worker.WorkerSessionIDHandler)
-	if !ok {
-		return
-	}
-	workerSID := handler.GetWorkerSessionID()
-	if workerSID == "" {
-		return
-	}
-	if err := b.sm.UpdateWorkerSessionID(context.Background(), sessionID, workerSID); err != nil {
-		b.log.Warn("bridge: failed to persist worker session ID", "session_id", sessionID, "worker_session_id", workerSID, "err", err)
-	} else {
-		b.log.Debug("bridge: persisted worker session ID", "session_id", sessionID, "worker_session_id", workerSID)
-	}
-}
-
-// forwardEvents proxies worker events to the hub with seq assignment.
-// EVT-004: if msgStore is configured, it appends to the event log on done events.
-// AEP-020: after the recv channel closes, calls Worker.Wait() to determine exit
-// code and sets DoneData.Success accordingly (non-zero exit = crash = success=false).
-func (b *Bridge) forwardEvents(w worker.Worker, sessionID string, opts forwardOpts) {
-	defer func() {
-		if r := recover(); r != nil {
-			b.log.Error("bridge: panic in forwardEvents", "session_id", sessionID, "panic", r, "stack", string(debug.Stack()))
-		}
-	}()
-	workerType := w.Type()
-	b.log.Debug("bridge: forwardEvents goroutine started", "session_id", sessionID, "worker_type", workerType, "resumed", opts.resumed)
-
-	// Cache session info for conversation store writes (avoids per-turn DB lookup).
-	var sessPlatform, sessOwner string
-	if b.convStore != nil && b.sm != nil {
-		if si, err := b.sm.Get(sessionID); err == nil {
-			sessPlatform = si.Platform
-			sessOwner = si.OwnerID
-		}
-	}
-	startTime := time.Now()
-	turnStartTime := startTime
-	firstEvent := true
-	doneReceived := false
-
-	// Capture reset generation at goroutine start. If a reset happens while
-	// this goroutine is running, the generation will differ when we check
-	// after the recv channel closes, and we exit cleanly without crash handling.
-	var myGen int64
-	if rg, ok := w.(resetGenerationer); ok {
-		myGen = rg.LoadResetGeneration()
-	}
-
-	// LLM retry: accumulate turn text and error data for retry detection.
-	var turnText strings.Builder
-	var lastError *events.ErrorData
-	var pendingError *events.Envelope // buffered error event; suppressed if retry triggers
-
-	// Turn timeout: kill worker if a single turn exceeds the configured duration.
-	// Timer resets on every received event; stops on done.
-	var turnTimerFired atomic.Bool
-	var turnTimer *time.Timer
-	if b.turnTimeout > 0 {
-		turnTimer = time.AfterFunc(b.turnTimeout, func() {
-			if !turnTimerFired.CompareAndSwap(false, true) {
-				return
-			}
-			b.log.Warn("bridge: turn timeout exceeded, terminating worker",
-				"session_id", sessionID, "worker_type", workerType, "turn_timeout", b.turnTimeout)
-			b.sendError(sessionID, events.ErrCodeTurnTimeout, "Turn exceeded %v time limit and was terminated.", b.turnTimeout)
-			_ = w.Terminate(context.Background())
-		})
-		defer turnTimer.Stop()
-	}
-
-	recvCh := w.Conn().Recv()
-
-	for env := range recvCh {
-
-		if env.Event.Type == events.Error {
-			b.log.Warn("bridge: received error from worker", "session_id", sessionID, "worker_type", workerType, "data", env.Event.Data)
-			if ed, ok := env.Event.Data.(events.ErrorData); ok {
-				lastError = &ed
-			}
-			if b.retryCtrl != nil {
-				cloned := events.Clone(env)
-				cloned.SessionID = sessionID
-				pendingError = cloned
-				continue
-			}
-		} else if b.log.Enabled(context.Background(), slog.LevelDebug) {
-			b.log.Debug("bridge: received event from worker", "session_id", sessionID, "worker_type", workerType, "event_type", env.Event.Type)
-		}
-
-		if firstEvent {
-			b.persistWorkerSessionID(w, sessionID)
-			firstEvent = false
-		}
-
-		if turnTimer != nil && !turnTimerFired.Load() {
-			turnTimer.Reset(b.turnTimeout)
-		}
-		if turnTimerFired.Load() {
-			continue
-		}
-
-		env = events.Clone(env)
-		env.SessionID = sessionID
-
-		if env.Event.Type == events.MessageDelta || env.Event.Type == events.Message {
-			if content := extractMessageContent(env); content != "" {
-				turnText.WriteString(content)
-			}
-		}
-
-		// Stats accumulation: track tool calls and merge per-turn stats on done.
-		switch env.Event.Type {
-		case events.ToolCall:
-			acc := b.getOrInitAccum(sessionID, "")
-			acc.ToolCallCount++
-			if tc, ok := asToolCallData(env.Event.Data); ok {
-				if acc.ToolNames == nil {
-					acc.ToolNames = make(map[string]int)
-				}
-				acc.ToolNames[tc.Name]++
-			}
-		case events.Done:
-			if turnTimer != nil {
-				turnTimer.Stop()
-			}
-			acc := b.getOrInitAccum(sessionID, opts.workDir)
-			if dd, ok := asDoneData(env.Event.Data); ok {
-				acc.mergePerTurnStats(dd)
-			}
-			acc.TurnCount++
-			acc.TurnDurationMs = time.Since(turnStartTime).Milliseconds()
-			acc.computePerTurnDeltas()
-
-			// Query precise context usage from worker via control channel.
-			// Silently falls back to aggregated Done stats on failure.
-			if cr, ok := w.(ControlRequester); ok {
-				ctrlCtx, ctrlCancel := context.WithTimeout(context.Background(), 5*time.Second)
-				if resp, err := cr.SendControlRequest(ctrlCtx, "get_context_usage", nil); err == nil {
-					if cu := events.MapContextUsageResponse(resp); cu.MaxTokens > 0 {
-						acc.mergeContextUsage(cu)
-					}
-				}
-				ctrlCancel()
-			}
-
-			b.injectSessionStats(env, acc)
-			if b.log.Enabled(context.Background(), slog.LevelDebug) {
-				b.log.Debug("bridge: turn completed",
-					"session_id", sessionID, "worker_type", workerType, "turn", acc.TurnCount,
-					"duration", time.Since(turnStartTime).Round(time.Millisecond),
-					"text_len", turnText.Len(), "tools", acc.ToolCallCount)
-			}
-		}
-
-		// UI Reconciliation (Fallback full message if silent dropped)
-		if env.Event.Type == events.Done {
-			doneReceived = true
-			b.resetCrashLoop(sessionID)
-			if b.hub.GetAndClearDropped(sessionID) {
-				b.log.Warn("bridge: handling dropped deltas before done", "session_id", sessionID, "worker_type", workerType)
-
-				if dataMap, ok := env.Event.Data.(map[string]any); ok {
-					if stats, ok := dataMap["stats"].(map[string]any); ok {
-						stats["dropped"] = true
-					} else {
-						dataMap["stats"] = map[string]any{"dropped": true}
-					}
-				} else if doneData, ok := env.Event.Data.(events.DoneData); ok {
-					doneData.Dropped = true
-					env.Event.Data = doneData
-				} else if doneDataPtr, ok := env.Event.Data.(*events.DoneData); ok {
-					doneDataPtr.Dropped = true
-					env.Event.Data = doneDataPtr
-				}
-			}
-		}
-
-		if err := b.hub.SendToSession(context.Background(), env); err != nil {
-			b.log.Warn("bridge: forward event failed", "err", err, "session_id", sessionID, "worker_type", workerType, "event_type", env.Event.Type)
-		}
-
-		b.captureEvent(sessionID, env.Seq, env.Event.Type, env.Event.Data)
-
-		// Flush buffered error on non-Done events (no retry decision possible yet).
-		if pendingError != nil && env.Event.Type != events.Done {
-			if err := b.hub.SendToSession(context.Background(), pendingError); err != nil {
-				b.log.Warn("bridge: forward buffered error failed", "err", err, "session_id", sessionID, "worker_type", workerType)
-			}
-			b.captureEvent(sessionID, pendingError.Seq, pendingError.Event.Type, pendingError.Event.Data)
-			pendingError = nil
-		}
-
-		// LLM retry: check after Done is forwarded.
-		if env.Event.Type == events.Done && b.retryCtrl != nil && (!opts.resumed || turnText.Len() > 0) {
-			if shouldRetry, attempt := b.retryCtrl.ShouldRetry(sessionID, lastError); shouldRetry {
-				pendingError = nil
-				b.autoRetry(context.Background(), w, sessionID, attempt)
-				turnText.Reset()
-				lastError = nil
-				continue
-			}
-			if pendingError != nil {
-				if err := b.hub.SendToSession(context.Background(), pendingError); err != nil {
-					b.log.Warn("bridge: forward buffered error failed", "err", err, "session_id", sessionID, "worker_type", workerType)
-				}
-				b.captureEvent(sessionID, pendingError.Seq, pendingError.Event.Type, pendingError.Event.Data)
-				pendingError = nil
-			}
-			b.retryCtrl.RecordSuccess(sessionID)
-			lastError = nil
-		}
-
-		// Record assistant response to conversation store after retry decision.
-		if env.Event.Type == events.Done && b.convStore != nil {
-			dd, _ := asDoneData(env.Event.Data)
-			acc := b.getOrInitAccum(sessionID, "")
-			success := dd.Success
-			_ = b.convStore.Append(context.Background(), &session.ConversationRecord{
-				SessionID:     sessionID,
-				Seq:           env.Seq,
-				Role:          session.RoleAssistant,
-				Content:       turnText.String(),
-				Platform:      sessPlatform,
-				UserID:        sessOwner,
-				Model:         acc.ModelName,
-				Success:       &success,
-				Source:        session.SourceNormal,
-				Tools:         acc.ToolNames,
-				ToolCallCount: acc.ToolCallCount,
-				TokensIn:      acc.PerTurnInput,
-				TokensOut:     acc.PerTurnOutput,
-				DurationMs:    acc.TurnDurationMs,
-				CostUSD:       acc.PerTurnCost,
-			})
-			acc.resetPerTurn()
-		}
-
-		if env.Event.Type == events.Done {
-			turnText.Reset()
-			turnStartTime = time.Now()
-		}
-	}
-
-	// Check reset generation: if a reset happened while this goroutine was
-	// running, the generation counter will differ from our captured value.
-	// This is race-free because the counter is monotonic — OLD forwardEvents
-	// always sees a different generation than what it captured at start.
-	if rg, ok := w.(resetGenerationer); ok && rg.LoadResetGeneration() != myGen {
-		b.log.Info("bridge: worker reset, old forwardEvents exiting", "session_id", sessionID, "worker_type", workerType, "my_gen", myGen, "cur_gen", rg.LoadResetGeneration())
-		return
-	}
-
-	// AEP-020: Worker.Recv() closed — get exit code to determine crash vs normal exit.
-	// Wrap Wait() with a timeout to avoid blocking the goroutine if the process
-	// is stuck in an unreported state (e.g. zombie or kernel-level hang).
-	// Use a longer timeout during shutdown to allow graceful termination.
-	waitTimeout := 2 * time.Second
-	if b.closed.Load() {
-		waitTimeout = 10 * time.Second
-	}
-	var exitCode int
-	ch := make(chan struct{})
-	go func() {
-		exitCode, _ = w.Wait()
-		close(ch)
-	}()
-	select {
-	case <-ch:
-		// Wait() returned; check exit code.
-	case <-time.After(waitTimeout):
-		// Force-kill to ensure the Wait() goroutine completes and doesn't leak.
-		b.log.Warn("bridge: Wait() timed out, force-killing", "session_id", sessionID, "worker_type", workerType)
-		_ = w.Kill()
-		<-ch // drain the goroutine
-	}
-
-	// Resume retry: if the resumed worker crashed quickly (within 15s), it likely
-	// failed to restore its conversation state. Retry resume once (limited by
-	// retryDepth) to recover, then fall back to fresh start if retry also fails.
-	// Skip during shutdown to avoid spawning workers that will be immediately killed.
-	// Skip for exit 143 (SIGTERM) — intentional termination by a new connection, not a crash.
-	fallbackAttempted := !b.closed.Load() && exitCode != 0 && exitCode != 143 && opts.resumed && opts.retryDepth < 2 && time.Since(startTime) < 15*time.Second
-	// Session files missing: resumed worker crashed immediately with no output.
-	// Skip the resume retry (guaranteed to fail again) and go directly to fresh start.
-	if fallbackAttempted && turnText.Len() == 0 && time.Since(startTime) < 5*time.Second {
-		b.log.Info("bridge: session files missing after resume, skipping retry",
-			"session_id", sessionID, "worker_type", workerType, "exit_code", exitCode)
-		opts.retryDepth = 1
-	}
-	if fallbackAttempted {
-		// Extract last input from dead worker's conn for re-delivery after fresh start.
-		// Fall back to inherited lastInput from previous retry goroutine when the
-		// retry worker never received any input (conn.LastInput is empty).
-		var lastInput string
-		if conn := w.Conn(); conn != nil {
-			if ir, ok := conn.(worker.InputRecoverer); ok {
-				lastInput = sanitizeLastInput(ir.LastInput())
-			}
-		}
-		if lastInput == "" {
-			lastInput = opts.lastInput
-		}
-		if b.attemptResumeFallback(fallbackParams{
-			sessionID:     sessionID,
-			workDir:       opts.workDir,
-			exitCode:      exitCode,
-			retryDepth:    opts.retryDepth,
-			workerType:    workerType,
-			lastInput:     lastInput,
-			crashedWorker: w,
-		}) {
-			return // new forwardEvents goroutine took over
-		}
-		// Fallback already cleaned up (DetachWorker + Transition); skip redundant cleanup below.
-	}
-
-	// During shutdown, skip crash detection — workers are SIGTERM'd by design.
-	// Just clean up without sending crash done events or incrementing crash metrics.
-	if b.closed.Load() {
-		b.cleanupCrashedWorker(sessionID, w)
-		return
-	}
-
-	// If session is already TERMINATED (e.g., handler client_kill), the handler
-	// already sent an error event. Skip sending redundant crash error — only
-	// detach and clean up.
-	if b.sm != nil {
-		si, smErr := b.sm.Get(sessionID)
-		if smErr == nil && si.State == events.StateTerminated {
-			b.log.Debug("bridge: session already terminated, skipping error for handler-killed worker", "session_id", sessionID, "worker_type", workerType)
-			if !fallbackAttempted {
-				b.cleanupCrashedWorker(sessionID, w)
-			}
-			return
-		}
-	}
-
-	// Send Error events for cleanup — platform adapters handle Error the same as Done
-	// (clear indicators, cancel interactions, close streaming cards) but without
-	// triggering turn summary, which requires _session data from a real worker Done.
-	if exitCode != 0 && exitCode != 143 && exitCode != -1 {
-		acc := b.getOrInitAccum(sessionID, "")
-		b.log.Warn("bridge: worker exited with non-zero code, sending crash error",
-			"session_id", sessionID, "worker_type", workerType, "exit_code", exitCode,
-			"duration", time.Since(startTime).Round(time.Millisecond), "turn_count", acc.TurnCount)
-		metrics.WorkerCrashesTotal.WithLabelValues(string(workerType), fmt.Sprintf("%d", exitCode)).Inc()
-		b.sendError(sessionID, events.ErrCodeWorkerCrash, "worker crashed (exit code %d)", exitCode)
-	} else if exitCode == 143 {
-		b.sendError(sessionID, events.ErrCodeSessionTerminated, "worker terminated (SIGTERM)")
-	} else if exitCode == -1 {
-		b.sendError(sessionID, events.ErrCodeSessionTerminated, "worker terminated (killed)")
-	} else if !doneReceived {
-		b.log.Debug("bridge: sending error for platform cleanup (no done received)", "session_id", sessionID, "worker_type", workerType)
-		b.sendError(sessionID, events.ErrCodeWorkerCrash, "worker exited without sending done")
-	}
-
-	// Record partial assistant response on crash/timeout (Path 3).
-	workerCrashed := (exitCode != 0 && exitCode != 143 && exitCode != -1) || (!doneReceived && turnTimerFired.Load())
-	if b.convStore != nil && turnText.Len() > 0 && workerCrashed {
-		acc := b.getOrInitAccum(sessionID, "")
-		src := session.SourceCrash
-		if turnTimerFired.Load() {
-			src = session.SourceTimeout
-		}
-		falseVal := false
-		_ = b.convStore.Append(context.Background(), &session.ConversationRecord{
-			SessionID:     sessionID,
-			Seq:           b.hub.NextSeq(sessionID),
-			Role:          session.RoleAssistant,
-			Content:       turnText.String(),
-			Platform:      sessPlatform,
-			UserID:        sessOwner,
-			Model:         acc.ModelName,
-			Success:       &falseVal,
-			Source:        src,
-			Tools:         acc.ToolNames,
-			ToolCallCount: acc.ToolCallCount,
-		})
-	}
-
-	// Clean up: detach the dead worker and transition session to TERMINATED
-	// so the next message triggers orphan resume instead of silently dropping input.
-	// Skip when attemptResumeFallback already performed cleanup.
-	if !fallbackAttempted {
-		b.cleanupCrashedWorker(sessionID, w)
-	}
-}
-
-// fallbackParams carries the context needed by attemptResumeFallback.
-type fallbackParams struct {
-	sessionID     string
-	workDir       string
-	exitCode      int
-	retryDepth    int
-	workerType    worker.WorkerType
-	lastInput     string
-	crashedWorker worker.Worker
-}
-
-// attemptResumeFallback handles a crashed resumed worker with a two-step strategy:
-//  1. retryDepth < 1: Retry resume once to preserve conversation history (transient failures).
-//  2. retryDepth >= 1: Fall back to fresh start — conversation data is permanently lost.
-//
-// Returns true if a new forwardEvents goroutine took over.
-func (b *Bridge) attemptResumeFallback(p fallbackParams) bool {
-	b.log.Warn("bridge: worker crashed shortly after resume",
-		"session_id", p.sessionID, "worker_type", p.workerType, "exit_code", p.exitCode, "retry_depth", p.retryDepth)
-
-	// Crash-loop protection: if this session has crashed crashLoopMax times within
-	// crashLoopWindow, stop retrying to prevent thread exhaustion that can kill
-	// the entire gateway process (pthread_create failed → SIGABRT).
-	if b.recordCrashLoop(p.sessionID) {
-		b.log.Error("bridge: crash loop detected, aborting retry to protect gateway",
-			"session_id", p.sessionID, "worker_type", p.workerType, "max", crashLoopMax, "window", crashLoopWindow)
-		b.sendError(p.sessionID, events.ErrCodeWorkerCrash, "Worker crashed %d times in %v. Stopping retry to protect gateway stability.", crashLoopMax, crashLoopWindow)
-		return false
-	}
-
-	// Clean up the crashed worker first.
-	b.cleanupCrashedWorker(p.sessionID, p.crashedWorker)
-
-	// Step 1: Retry resume once for transient failures (e.g., file lock, timing).
-	if p.retryDepth == 0 {
-		if err := b.resumeWithOpts(context.Background(), p.sessionID, p.workDir, forwardOpts{resumed: true, workDir: p.workDir, retryDepth: p.retryDepth + 1, lastInput: p.lastInput}); err != nil {
-			b.log.Warn("bridge: resume retry failed synchronously, falling back to fresh start", "session_id", p.sessionID, "worker_type", p.workerType, "err", err)
-			// Synchronous failure — fall through to fresh start below.
-		} else {
-			b.log.Info("bridge: resume retry succeeded", "session_id", p.sessionID, "worker_type", p.workerType)
-			b.sendError(p.sessionID, events.ErrCodeResumeRetry, "Worker crashed after resume (exit %d), retried resume to preserve conversation.", p.exitCode)
-			return true
-		}
-	}
-
-	// Step 2: Resume retry also failed or retryDepth exhausted — start fresh worker.
-	// Conversation data is permanently lost (e.g., "No conversation found").
-	b.log.Info("bridge: starting fresh worker after failed resume", "session_id", p.sessionID, "worker_type", p.workerType)
-
-	si, err := b.sm.Get(p.sessionID)
-	if err != nil {
-		b.log.Error("bridge: session not found for fresh start fallback", "session_id", p.sessionID, "err", err)
-		return false
-	}
-
-	workerInfo := worker.SessionInfo{
-		SessionID:       si.ID,
-		UserID:          si.UserID,
-		AllowedTools:    si.AllowedTools,
-		WorkerSessionID: si.WorkerSessionID,
-		ProjectDir:      p.workDir,
-		ConfigEnv:       b.workerEnv,
-		ConfigWhitelist: b.workerEnvWhitelist,
-	}
-
-	// Inject Slack context for CLI subcommand auto-resolution (same as StartSession).
-	if si.PlatformKey != nil {
-		if chID, ok := si.PlatformKey["channel_id"]; ok && chID != "" {
-			if workerInfo.Env == nil {
-				workerInfo.Env = make(map[string]string)
-			}
-			workerInfo.Env["HOTPLEX_SLACK_CHANNEL_ID"] = chID
-			if threadTS, ok := si.PlatformKey["thread_ts"]; ok && threadTS != "" {
-				workerInfo.Env["HOTPLEX_SLACK_THREAD_TS"] = threadTS
-			}
-		}
-	}
-
-	w, err := b.createAndLaunchWorker(workerLaunchParams{
-		ctx:        context.Background(),
-		wt:         si.WorkerType,
-		workerInfo: workerInfo,
-		platform:   si.Platform,
-		botID:      si.BotID,
-	},
-		func(ctx context.Context, w worker.Worker, info worker.SessionInfo) error {
-			if err := b.sm.Transition(ctx, p.sessionID, events.StateRunning); err != nil {
-				b.log.Warn("bridge: transition to running for fresh start", "session_id", p.sessionID, "err", err)
-			}
-			if err := w.Start(ctx, info); err != nil {
-				b.log.Warn("bridge: fresh worker start failed", "session_id", p.sessionID, "err", err)
-				return err
-			}
-			return nil
-		},
-		func(_ worker.Worker, err error) {
-			b.log.Error("bridge: attach worker for fresh start", "session_id", p.sessionID, "err", err)
-		},
-	)
-	if err != nil {
-		return false
-	}
-
-	// Re-deliver the original input that was lost when the first worker crashed.
-	if p.lastInput != "" {
-		b.log.Info("bridge: re-delivering input to fresh worker", "session_id", p.sessionID, "content_len", len(p.lastInput))
-		if err := w.Input(context.Background(), p.lastInput, nil); err != nil {
-			b.log.Warn("bridge: input re-delivery failed", "session_id", p.sessionID, "err", err)
-		} else if b.convStore != nil {
-			_ = b.convStore.Append(context.Background(), &session.ConversationRecord{
-				SessionID: p.sessionID,
-				Seq:       b.hub.NextSeq(p.sessionID),
-				Role:      session.RoleUser,
-				Content:   p.lastInput,
-				Platform:  si.Platform,
-				UserID:    si.OwnerID,
-				Source:    session.SourceFreshStart,
-			})
-		}
-	}
-
-	b.log.Info("bridge: fresh worker started after resume failure", "session_id", p.sessionID, "worker_type", p.workerType)
-	notifyMsg := buildNotifyEnvelope(p.sessionID,
-		"🔄 会话已重新启动，上下文已重置。",
-		b.hub.NextSeq(p.sessionID))
-	_ = b.hub.SendToSession(context.Background(), notifyMsg)
-	return true
-}
-
-// cleanupCrashedWorker detaches the dead worker and transitions the session to TERMINATED
-// so the next message triggers orphan resume instead of silently dropping input.
-// Uses CAS via crashedWorker to avoid detaching a worker that was already replaced
-// by a concurrent ResumeSession or attemptResumeFallback.
-func (b *Bridge) cleanupCrashedWorker(sessionID string, crashedWorker worker.Worker) {
-	// Read turn_count without deleting — the accumulator is cleaned up when the
-	// session is physically deleted, not here. This avoids reading TurnCount=0
-	// after resetPerTurn() has already cleared it at done.
-	acc := b.getOrInitAccum(sessionID, "")
-	b.log.Debug("bridge: cleaning up crashed worker", "session_id", sessionID, "turn_count", acc.TurnCount)
-	if b.sm == nil {
-		return
-	}
-	if crashedWorker != nil {
-		if !b.sm.DetachWorkerIf(sessionID, crashedWorker) {
-			b.log.Debug("bridge: crashed worker already replaced, skipping cleanup", "session_id", sessionID)
-			return
-		}
-	} else {
-		b.sm.DetachWorker(sessionID)
-	}
-	if err := b.sm.Transition(context.Background(), sessionID, events.StateTerminated); err != nil {
-		b.log.Debug("bridge: transition to terminated after worker exit", "session_id", sessionID, "err", err)
-	}
-}
 
 // StartPlatformSession creates a session for a platform message if it doesn't already exist.
 // Implements messaging.SessionStarter. Idempotent: returns nil if session exists with a live worker.
@@ -1117,156 +500,9 @@ func (b *Bridge) Shutdown(ctx context.Context) {
 	}
 }
 
-// CancelRetry cancels any pending auto-retry for a session.
-// Called by handler when a user sends a new input.
-func (b *Bridge) CancelRetry(sessionID string) {
-	b.retryCancelMu.Lock()
-	defer b.retryCancelMu.Unlock()
-	if ch, ok := b.retryCancel[sessionID]; ok {
-		close(ch)
-		delete(b.retryCancel, sessionID)
-	}
-}
-
-// injectAgentConfig loads agent config files and injects the unified system
-// prompt into session info. A no-op when config dir is empty or agent config
-// is not configured.
-func (b *Bridge) injectAgentConfig(info *worker.SessionInfo, platform, botID string) {
-	if b.agentConfigDir == "" {
-		return
-	}
-	b.log.Debug("bridge: loading agent config", "dir", b.agentConfigDir, "platform", platform, "bot_id", botID)
-	configs, err := agentconfig.Load(b.agentConfigDir, platform, botID)
-	if err != nil {
-		if strings.Contains(err.Error(), "invalid botID") {
-			b.log.Error("bridge: agent config rejected botID",
-				"dir", b.agentConfigDir, "platform", platform, "bot_id", botID, "err", err)
-		} else {
-			b.log.Warn("bridge: agent config load failed",
-				"dir", b.agentConfigDir, "platform", platform, "bot_id", botID, "err", err)
-		}
-		return
-	}
-	if configs.IsEmpty() {
-		b.log.Warn("bridge: agent config empty, no files found",
-			"dir", b.agentConfigDir, "platform", platform, "bot_id", botID)
-		return
-	}
-
-	if prompt := agentconfig.BuildSystemPrompt(configs); prompt != "" {
-		info.SystemPrompt = prompt
-		b.log.Info("bridge: agent config injected", "prompt_len", len(prompt), "platform", platform, "bot_id", botID)
-	} else {
-		b.log.Debug("bridge: agent config loaded but prompt empty", "platform", platform, "bot_id", botID)
-	}
-}
-
-// autoRetry performs exponential backoff then sends the retry input to the worker.
-func (b *Bridge) autoRetry(ctx context.Context, w worker.Worker, sessionID string, attempt int) {
-	delay := b.retryCtrl.Delay(attempt)
-
-	// Notify user if enabled.
-	if b.retryCtrl.ShouldNotify() {
-		msg := b.retryCtrl.NotifyMessage(attempt)
-		notifyEnv := buildNotifyEnvelope(sessionID, msg, b.hub.NextSeq(sessionID))
-		_ = b.hub.SendToSession(ctx, notifyEnv)
-	}
-
-	// Register cancel channel.
-	cancelCh := make(chan struct{})
-	b.retryCancelMu.Lock()
-	b.retryCancel[sessionID] = cancelCh
-	b.retryCancelMu.Unlock()
-	defer func() {
-		b.retryCancelMu.Lock()
-		delete(b.retryCancel, sessionID)
-		b.retryCancelMu.Unlock()
-	}()
-
-	// Wait with backoff, respecting cancellation.
-	timer := time.NewTimer(delay)
-	defer timer.Stop()
-	select {
-	case <-ctx.Done():
-		return
-	case <-cancelCh:
-		b.log.Info("bridge: auto-retry cancelled by user input", "session_id", sessionID)
-		return
-	case <-timer.C:
-	}
-
-	// Send retry input to worker.
-	b.log.Info("bridge: auto-retry sending input", "session_id", sessionID, "attempt", attempt)
-	if err := w.Input(ctx, b.retryCtrl.RetryInput(), nil); err != nil {
-		b.log.Warn("bridge: auto-retry input failed", "session_id", sessionID, "err", err)
-	}
-}
-
-func (b *Bridge) captureEvent(sessionID string, seq int64, eventType events.Kind, data any) {
-	if b.collector == nil {
-		return
-	}
-	ed, err := json.Marshal(data)
-	if err != nil {
-		return
-	}
-	b.collector.Capture(sessionID, seq, eventType, ed, "outbound")
-}
-
-// extractMessageContent extracts text content from a message or message_delta event.
-func extractMessageContent(env *events.Envelope) string {
-	switch env.Event.Type {
-	case events.Message, events.MessageDelta:
-		if m, ok := env.Event.Data.(map[string]any); ok {
-			if content, ok := m["content"].(string); ok {
-				return content
-			}
-		}
-	}
-	return ""
-}
-
 // buildNotifyEnvelope creates a synthetic Message event for user notifications.
 func buildNotifyEnvelope(sessionID, msg string, seq int64) *events.Envelope {
 	return events.NewEnvelope(aep.NewID(), sessionID, seq, events.Message, map[string]any{"content": msg})
-}
-
-// getOrInitAccum returns the session accumulator, creating one if needed.
-func (b *Bridge) getOrInitAccum(sessionID, workDir string) *sessionAccumulator {
-	b.accumMu.Lock()
-	defer b.accumMu.Unlock()
-	if acc, ok := b.accum[sessionID]; ok {
-		return acc
-	}
-	acc := &sessionAccumulator{StartedAt: time.Now()}
-	if workDir != "" {
-		acc.WorkDir = workDir
-		acc.GitBranch = gitBranchOf(workDir)
-	}
-	b.accum[sessionID] = acc
-	return acc
-}
-
-// injectSessionStats merges the accumulator snapshot into DoneData.Stats["_session"].
-// Handles both typed DoneData and map[string]any (from events.Clone JSON round-tripping).
-func (b *Bridge) injectSessionStats(env *events.Envelope, acc *sessionAccumulator) {
-	dd, ok := asDoneData(env.Event.Data)
-	if !ok {
-		return
-	}
-	if dd.Stats == nil {
-		dd.Stats = make(map[string]any)
-	}
-	dd.Stats["_session"] = acc.snapshot()
-
-	// Write back: preserve the original representation (map stays map, struct stays struct).
-	switch env.Event.Data.(type) {
-	case map[string]any:
-		raw, _ := json.Marshal(dd)
-		_ = json.Unmarshal(raw, &env.Event.Data)
-	default:
-		env.Event.Data = dd
-	}
 }
 
 // sanitizeLastInput filters control-like text from lastInput before re-delivery
@@ -1297,66 +533,10 @@ func sanitizeLastInput(input string) string {
 	return strings.Join(filtered, "\n")
 }
 
-// recordCrashLoop tracks consecutive crashes per session. Returns true if the
-// session has exceeded crashLoopMax crashes within crashLoopWindow, indicating
-// a crash loop that should abort retries to protect gateway stability.
-func (b *Bridge) recordCrashLoop(sessionID string) bool {
-	b.crashTrackerMu.Lock()
-	defer b.crashTrackerMu.Unlock()
-
-	h, ok := b.crashTracker[sessionID]
-	if !ok || time.Since(h.firstSeen) > crashLoopWindow {
-		h = &crashHistory{firstSeen: time.Now()}
-		b.crashTracker[sessionID] = h
-	}
-	h.count++
-
-	return h.count > crashLoopMax
-}
-
-// resetCrashLoop clears the crash history for a session on successful completion.
-func (b *Bridge) resetCrashLoop(sessionID string) {
-	b.crashTrackerMu.Lock()
-	defer b.crashTrackerMu.Unlock()
-	delete(b.crashTracker, sessionID)
-}
-
 func (b *Bridge) sendError(sessionID string, code events.ErrorCode, format string, args ...any) {
 	env := events.NewEnvelope(aep.NewID(), sessionID, b.hub.NextSeq(sessionID), events.Error, events.ErrorData{
 		Code:    code,
 		Message: fmt.Sprintf(format, args...),
 	})
 	_ = b.hub.SendToSession(context.Background(), env)
-}
-
-// gitAvailable is checked once via sync.Once. If git is not on PATH,
-// branch detection is skipped entirely for all sessions.
-var (
-	gitAvailable bool
-	gitOnce      sync.Once
-)
-
-func checkGitAvailable() bool {
-	gitOnce.Do(func() {
-		_, err := exec.LookPath("git")
-		gitAvailable = err == nil
-	})
-	return gitAvailable
-}
-
-// gitBranchOf returns the current git branch name for the given directory, or empty string.
-// Best-effort: 2s timeout, skips if git is not installed, errors silently ignored.
-func gitBranchOf(dir string) string {
-	if !checkGitAvailable() {
-		return ""
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	cmd := exec.CommandContext(ctx, "git", "rev-parse", "--abbrev-ref", "HEAD")
-	cmd.Dir = dir
-	out, err := cmd.Output()
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(string(out))
 }
